# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API",
    "version": "1.0.0"
  },
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "modelo": {
          "type": "string"
        },
        "potencia": {
          "type": "integer"
        },
        "precio": {
          "type": "integer"
        }
      },
      "required": ["modelo", "potencia", "precio"]
    },
    "Concesionario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "listado_coches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coche"
          }
        }
      },
      "required": ["nombre", "direccion"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Fetch concesionarios",
        "operationId": "getConcesionarios",
        "responses": {
          "200": {
            "description": "List of concesionarios.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionario"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create concesionario",
        "operationId": "createConcesionario",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "concesionario",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Get a concesionario by id",
        "operationId": "getConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch concesionario.",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        }
      },
      "put": {
        "summary": "Update concesionario",
        "operationId": "updateConcesionario",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "concesionario",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      },
      "delete": {
        "summary": "Delete concesionario",
        "operationId": "deleteConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "{message: 'ok'}"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Fetch coches for a concesionario",
        "operationId": "getCoches",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of coches in concesionario.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Coche"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create coche for a concesionario",
        "operationId": "createCoche",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "coche",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche created",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        }
      }
    },
    "/concesionarios/{id}/coches/{cocheId}": {
      "get": {
        "summary": "Get a coche by id",
        "operationId": "getCoche",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch coche details.",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        }
      },
      "put": {
        "summary": "Update coche by id",
        "operationId": "updateCoche",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "coche",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche updated"
          }
        }
      },
      "delete": {
        "summary": "Delete coche by id",
        "operationId": "deleteCoche",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Coche deleted"
          }
        }
      }
    }
  }
}